file management + start input
foundation for MNIST database intake
changed where the code was located
